move lib/templates-helper to TemplatesCommand
